dashboard index page-> rules added

+++ database/migrations/2020_08_17_034403_create_clubs_table.php
@@ -17,8 +17,8 @@ class CreateClubsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('is_active')->default(1);
-            $table->float('balance')->nullable();
-            $table->float('lock_balance')->nullable();
+            $table->double('balance')->nullable();
+            $table->double('lock_balance')->nullable();
             $table->unsignedBigInteger('member')->nullable();
             // $table->string('email');
             $table->float('commission')->nullable();
